refactor: modularize vector search pipeline
