+++ src/helpers/vectorSearch/context.js
@@ -1,0 +1,87 @@
+const CHUNK_SIZE = 1500;
+const OVERLAP = 100;
+
+function splitIntoSections(lines) {
+  const heading = /^(#{1,6})\s+/;
+  const sections = [];
+  let i = 0;
+  while (i < lines.length && !heading.test(lines[i])) i++;
+  if (i > 0) {
+    const pre = lines.slice(0, i).join("\n").trim();
+    if (pre) sections.push(pre);
+  }
+  collectHeadingSections(lines, i, sections);
+  return sections;
+}
+
+function collectHeadingSections(lines, startIdx, out) {
+  const heading = /^(#{1,6})\s+/;
+  for (let idx = startIdx; idx < lines.length; idx++) {
+    const match = heading.exec(lines[idx]);
+    if (!match) continue;
+    const level = match[1].length;
+    let j = idx + 1;
+    while (j < lines.length) {
+      const next = heading.exec(lines[j]);
+      if (next && next[1].length <= level) break;
+      j++;
+    }
+    const section = lines.slice(idx, j).join("\n").trim();
+    if (section) out.push(section);
+  }
+}
+
+function chunkSection(section) {
+  if (section.length <= CHUNK_SIZE) return [section];
+  const chunks = [];
+  for (let start = 0; start < section.length; start += CHUNK_SIZE - OVERLAP) {
+    chunks.push(section.slice(start, start + CHUNK_SIZE));
+  }
+  return chunks;
+}
+
+export function chunkMarkdown(text) {
+  const lines = text.split(/\r?\n/);
+  const sections = splitIntoSections(lines);
+  const chunks = [];
+  for (const section of sections) {
+    chunks.push(...chunkSection(section));
+  }
+  return chunks;
+}
+
+/**
+ * Fetch neighboring context chunks for a given embedding id.
+ *
+ * @pseudocode
+ * 1. Validate that `id` matches the `filename-chunk-N` pattern.
+ *    - Return an empty array for invalid ids.
+ * 2. Build a URL to the markdown file using the filename.
+ * 3. Fetch the markdown text with `fetch` and split it using `chunkMarkdown`.
+ * 4. Determine the slice of chunks around the requested index based on `radius`.
+ * 5. Return the selected chunk texts.
+ *
+ * @param {string} id - Entry identifier like `foo.md-chunk-3`.
+ * @param {number} [radius=1] - Number of neighboring chunks to include.
+ * @returns {Promise<string[]>} Array of surrounding chunk strings.
+ */
+export async function fetchContextById(id, radius = 1) {
+  const match = /^([^\s]+\.md)-chunk-(\d+)$/.exec(id);
+  if (!match) return [];
+  const [, filename, num] = match;
+  const index = Number(num) - 1;
+  try {
+    const url = new URL(`../../design/productRequirementsDocuments/${filename}`, import.meta.url)
+      .href;
+    const res = await fetch(url);
+    if (!res.ok) return [];
+    const text = await res.text();
+    const chunks = chunkMarkdown(text);
+    const start = Math.max(0, index - radius);
+    const end = Math.min(chunks.length, index + radius + 1);
+    return chunks.slice(start, end);
+  } catch (err) {
+    console.error(`Failed to load context from ${filename}`, err);
+    return [];
+  }
+}
