+++ tests/helpers/vectorSearch.test.js
@@ -27,7 +27,7 @@ const sample = [
 
 describe("vectorSearch", () => {
   it("loads embeddings once and caches", async () => {
-    const { loadEmbeddings } = await import("../../src/helpers/vectorSearch.js");
+    const { loadEmbeddings } = await import("../../src/helpers/vectorSearch/loader.js");
     const first = await loadEmbeddings();
     const second = await loadEmbeddings();
     expect(first).toEqual(sample);
@@ -38,20 +38,20 @@ describe("vectorSearch", () => {
   it("returns null when loading fails", async () => {
     const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
     fetchJsonMock.mockRejectedValueOnce(new Error("fail"));
-    const { loadEmbeddings } = await import("../../src/helpers/vectorSearch.js");
+    const { loadEmbeddings } = await import("../../src/helpers/vectorSearch/loader.js");
     const result = await loadEmbeddings();
     expect(result).toBeNull();
     errorSpy.mockRestore();
   });
 
   it("computes cosine similarity", async () => {
-    const { cosineSimilarity } = await import("../../src/helpers/vectorSearch.js");
+    const { cosineSimilarity } = await import("../../src/helpers/vectorSearch/scorer.js");
     expect(cosineSimilarity([1, 0], [1, 0])).toBeCloseTo(1);
     expect(cosineSimilarity([1, 0], [0, 1])).toBeCloseTo(0);
   });
 
   it("finds top matches", async () => {
-    const { findMatches } = await import("../../src/helpers/vectorSearch.js");
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 0], 1);
     expect(res.length).toBe(1);
     expect(res[0].id).toBe("a");
@@ -59,22 +59,35 @@ describe("vectorSearch", () => {
   });
 
   it("filters by tags", async () => {
-    const { findMatches } = await import("../../src/helpers/vectorSearch.js");
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 0], 1, ["foo"]);
     expect(res[0].id).toBe("a");
     const other = await findMatches([1, 0], 1, ["bar"]);
     expect(other[0].id).toBe("b");
   });
 
+  it("returns empty array when tags exclude all entries", async () => {
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
+    const res = await findMatches([1, 0], 5, ["baz"]);
+    expect(res).toEqual([]);
+  });
+
   it("boosts results containing exact query terms", async () => {
-    const { findMatches } = await import("../../src/helpers/vectorSearch.js");
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 1], 2, [], "B");
     expect(res[0].id).toBe("b");
   });
 
   it("returns empty array for dimension mismatch", async () => {
-    const { findMatches } = await import("../../src/helpers/vectorSearch.js");
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 0, 0], 1);
+    expect(res).toEqual([]);
+  });
+
+  it("returns empty array when embeddings dataset is empty", async () => {
+    fetchJsonMock.mockResolvedValueOnce([]);
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
+    const res = await findMatches([1, 0], 1);
     expect(res).toEqual([]);
   });
 
@@ -88,7 +101,7 @@ describe("vectorSearch", () => {
     ];
     fetchJsonMock.mockResolvedValueOnce(malformed);
     const warn = vi.spyOn(console, "warn").mockImplementation(() => {});
-    const { findMatches } = await import("../../src/helpers/vectorSearch.js");
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 0], 5);
     expect(res.map((r) => r.id)).toEqual(["a", "b"]);
     expect(warn).toHaveBeenCalledTimes(3);
@@ -98,7 +111,7 @@ describe("vectorSearch", () => {
   it("returns null when embeddings fail to load", async () => {
     const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
     fetchJsonMock.mockRejectedValueOnce(new Error("fail"));
-    const { findMatches } = await import("../../src/helpers/vectorSearch.js");
+    const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 0], 1);
     expect(res).toBeNull();
     errorSpy.mockRestore();
@@ -110,7 +123,7 @@ describe("vectorSearch", () => {
     const seg3 = "z".repeat(1600);
     const md = `${seg1}\n\n${seg2}\n\n${seg3}`;
     global.fetch = vi.fn().mockResolvedValue({ ok: true, text: async () => md });
-    const { fetchContextById } = await import("../../src/helpers/vectorSearch.js");
+    const { fetchContextById } = await import("../../src/helpers/vectorSearch/context.js");
     const result = await fetchContextById("doc.md-chunk-3", 1);
     expect(global.fetch).toHaveBeenCalled();
     expect(result).toHaveLength(3);
@@ -121,13 +134,13 @@ describe("vectorSearch", () => {
 
   it("chunks markdown by heading", async () => {
     const md = "## A\nText A\n### B\nText B\n## C\nText C";
-    const { chunkMarkdown } = await import("../../src/helpers/vectorSearch.js");
+    const { chunkMarkdown } = await import("../../src/helpers/vectorSearch/context.js");
     const chunks = chunkMarkdown(md);
     expect(chunks).toEqual(["## A\nText A\n### B\nText B", "### B\nText B", "## C\nText C"]);
   });
 
   it("returns empty array for invalid id", async () => {
-    const { fetchContextById } = await import("../../src/helpers/vectorSearch.js");
+    const { fetchContextById } = await import("../../src/helpers/vectorSearch/context.js");
     const res = await fetchContextById("badid");
     expect(res).toEqual([]);
   });
